backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.6
kafka-python==2.0.2
boto3==1.34.14
Pillow==10.2.0
reportlab==4.0.8
openpyxl==3.1.2
django-storages==1.14.2
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
drf-spectacular==0.27.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
